New column GENERATED_PAYMENTS_DATE for Period. New services.

+++ src/java/com/idega/block/finance/dao/PriceDAO.java
@@ -12,5 +12,5 @@ public interface PriceDAO extends GenericDao {
 	public static final String BEAN_NAME = "priceDao";
 
 	public Price getById(Long priceId);
-	public List<Price> getAllByPeriodId(Long periodId);
+	public List<Price> getAllByPeriodId(Integer periodId);
 }
